Mouse location tests
