+++ src/content.js
@@ -1,0 +1,19 @@
+const tasks = [
+    {
+      description: "Task 2",
+      completed: false,
+      index:1
+    },
+    {
+      description: "Task 1",
+      completed: false,
+      index:0
+    },
+    {
+      description: "Task 3",
+      completed: false,
+      index:2
+    }
+  ];
+
+export {tasks};
